Rewrite KV tests using proxy
Previously, we were only testing the `KVGateway`. In preparation for
porting the KV implementation to a worker, rewrite the tests to test
the full KV flow, including the `KVRouter`.
